[1] add authentication

+++ src/main/java/com/project/shopapp/dtos/OrderDetailDTO.java
@@ -25,7 +25,7 @@ public class OrderDetailDTO {
 
     @Min(value = 0, message = "Total money must be >= 0!")
     @JsonProperty("total_money")
-    private long totalMoney;
+    private Float totalMoney;
 
     private String color;
 }
